// lib/wordpress.d.ts
export interface WPCommentContent {
	rendered: string;
}

export interface Comment {
	id: number;
	post: number;
	parent: number;
	author: number; // 0 for guest
	author_name: string;
	author_url?: string;
	author_avatar_urls?: Record<string, string>;
	date: string;
	date_gmt: string;
	content: WPCommentContent;
	link: string;
	status?: string;
	type?: string;
}